Perf 7. New parser based on span for BlockGroup

+++ GeometryDashAPI/Levels/GameObjects/Default/Block.cs
@@ -12,7 +12,7 @@ namespace GeometryDashAPI.Levels.GameObjects.Default
         [GameProperty("3", 0, true)] public float PositionY { get; set; }
         [GameProperty("4", false)] public bool HorizontalReflection { get; set; }
         [GameProperty("5", false)] public bool VerticalReflection { get; set; }
-        [GameProperty("6", (int)0)] public int Rotation { get; set; }
+        [GameProperty("6", 0)] public int Rotation { get; set; }
         public bool Glow
         {
             get => !glow;
